# specs/001-product-manager-agent/contracts/api.yaml
openapi: 3.0.3
info:
  title: 产品经理 Agent API
  version: 1.0.0
  description: 智能产品需求分析与PRD生成助手的API契约
servers:
  - url: /api/v1
paths:
...
  /conversations/{conversation_id}/prd:
    get:
      summary: 获取会话的所有PRD版本
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PRD版本列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PRDDocument'
    
    post:
      summary: 生成新PRD版本
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sections:
                  type: array
                  items:
                    type: string
                  description: 指定要生成的章节（不传则生成完整PRD）
      responses:
        '201':
          description: PRD生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PRDDocument'
  /conversations/{conversation_id}/prd/{version_number}:
    get:
      summary: 获取指定版本的PRD
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version_number
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PRD详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PRDDocument'
    
    put:
      summary: 更新PRD状态
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version_number
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, confirmed, archived]
      responses:
        '200':
          description: 状态更新成功
  /conversations/{conversation_id}/prd/compare:
    post:
      summary: 对比两个PRD版本
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_version:
                  type: integer
                to_version:
                  type: integer
      responses:
        '200':
          description: 版本对比结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  from_version:
                    type: integer
                  to_version:
                    type: integer
                  diff:
                    type: string
                    description: diff格式的差异内容
  /conversations/{conversation_id}/prd/export:
    post:
      summary: 导出PRD
      tags:
        - PRD管理
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version_number:
                  type: integer
                format:
                  type: string
                  enum: [markdown, pdf, word]
                  description: 导出格式
      responses:
        '200':
          description: 导出成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  download_url:
                    type: string
                    format: uri

  # ========== 竞品分析 ==========
...
components:
  schemas:
...
    PRDDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        version_number:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [draft, confirmed, archived]
        created_at:
          type: string
          format: date-time
...